IKASAN-2135 final changes for job lock management work.

+++ ikasaneip/ootb/rest/scheduler-agent-rest-module/src/main/java/org/ikasan/job/orchestration/model/context/JobLockImpl.java
@@ -9,7 +9,7 @@ import org.ikasan.spec.scheduled.job.model.SchedulerJob;
 public class JobLockImpl implements JobLock {
 
     private String name;
-    private long lockCount = 1;
+    private int lockCount = 1;
     private Map<String, List<SchedulerJob>> jobs;
 
     @Override
@@ -23,12 +23,12 @@ public class JobLockImpl implements JobLock {
     }
 
     @Override
-    public void setLockCount(long lockCount) {
+    public void setLockCount(int lockCount) {
         this.lockCount = lockCount;
     }
 
     @Override
-    public long getLockCount() {
+    public int getLockCount() {
         return lockCount;
     }
